+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file starting fro
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 -----
+## [1.19.0] - 2020-02-29
+
+**Changed:**
+
+- Add gear.App.ListenWithContext method.
+- Add gear.ContextWithSignal function.
+- Add logging.SetLoggerLevel function.
+
 ## [1.18.0] - 2020-02-23
 
 **Changed:**